+++ backend/Dockerfile
@@ -8,7 +8,7 @@ COPY go.mod go.sum ./
 
 RUN go mod download
 
-COPY src/*.go ./
+COPY src/* ./
 
 RUN CGO_ENABLED=0 GOOS=linux go build -o ./bin/sidejob-api
