+++ backend/src/controllers/jobsController.go
@@ -4,37 +4,6 @@ import (
     "github.com/gin-gonic/gin"
 )
 
-func getUsers(c *gin.Context) {
-
-}
-
-func getUserByID(c *gin.Context) {
-
-}
-
-func getJobsForUser(c *gin.Context) {
-
-}
-
-func getJobForUser(c *gin.Context) {
-
-}
-
-func postUser(c *gin.Context) {
-
-}
-
-func postJobForUser(c *gin.Context) {
-
-}
-
-func patchUser(c *gin.Context) {
-
-}
-
-func patchJobForUser(c *gin.Context) {
-
-}
 
 func getJobs(c *gin.Context) {
