+++ backend/src/models/jobModel.go
@@ -4,20 +4,11 @@ import (
   "gorm.io/gorm"
 )
 
-type User struct {
-  gorm.Model
-  ID       uint
-  Name     string
-  Email    string `gorm:"type:varchar(100);uniqueIndex"`
-  Password string
-  Jobs     []Job
-}
-
 type Job struct {
+  gorm.Model
   ID          uint
   Name        string
   Address     string
   Description string
   UserID      uint
 }
-
