+++ backend/src/models/userModel.go
@@ -1,0 +1,14 @@
+package main
+
+import (
+  "gorm.io/gorm"
+)
+
+type User struct {
+  gorm.Model
+  ID       uint
+  Name     string
+  Email    string `gorm:"type:varchar(100);uniqueIndex"`
+  Password string
+  Jobs     []Job
+}
